fix eval of global abstract locations
