+++ src/cwe_checker_lib/src/analysis/pointer_inference/state/tests/mod.rs
@@ -3,6 +3,7 @@ use super::*;
 use crate::analysis::pointer_inference::object::*;
 use crate::{bitvec, def, expr, variable};
 
+mod access_handling;
 mod specialized_expressions;
 
 fn bv(value: i64) -> ValueDomain {
@@ -78,58 +79,6 @@ fn state() {
     assert_eq!(state.memory.get_num_objects(), 3);
     state.remove_unreferenced_objects();
     assert_eq!(state.memory.get_num_objects(), 2);
-}
-
-#[test]
-fn handle_store() {
-    let global_memory = RuntimeMemoryImage::mock();
-    let mut state = State::new(&variable!("RSP:8"), Tid::new("time0"), BTreeSet::new());
-    let stack_id = new_id("time0", "RSP");
-    assert_eq!(
-        state.eval(&expr!("RSP:8")),
-        Data::from_target(stack_id.clone(), bv(0))
-    );
-
-    state.handle_register_assign(&variable!("RSP:8"), &expr!("RSP:8 - 32:8"));
-    assert_eq!(
-        state.eval(&expr!("RSP:8")),
-        Data::from_target(stack_id.clone(), bv(-32))
-    );
-    state.handle_register_assign(&variable!("RSP:8"), &expr!("RSP:8 + -8:8"));
-    assert_eq!(
-        state.eval(&expr!("RSP:8")),
-        Data::from_target(stack_id.clone(), bv(-40))
-    );
-
-    state
-        .handle_store(&expr!("RSP:8 + 8:8"), &expr!("1:8"), &global_memory)
-        .unwrap();
-    state
-        .handle_store(&expr!("RSP:8 - 8:8"), &expr!("2:8"), &global_memory)
-        .unwrap();
-    state
-        .handle_store(&expr!("RSP:8 + -16:8"), &expr!("3:8"), &global_memory)
-        .unwrap();
-    state.handle_register_assign(&variable!("RSP:8"), &expr!("RSP:8 - 4:8"));
-
-    assert_eq!(
-        state
-            .load_value(&expr!("RSP:8 + 12:8"), ByteSize::new(8), &global_memory)
-            .unwrap(),
-        bv(1).into()
-    );
-    assert_eq!(
-        state
-            .load_value(&expr!("RSP:8 - 4:8"), ByteSize::new(8), &global_memory)
-            .unwrap(),
-        bv(2).into()
-    );
-    assert_eq!(
-        state
-            .load_value(&expr!("RSP:8 + -12:8"), ByteSize::new(8), &global_memory)
-            .unwrap(),
-        bv(3).into()
-    );
 }
 
 #[test]
@@ -224,61 +173,6 @@ fn reachable_ids_under_and_overapproximation() {
             .into_iter()
             .collect()
     );
-}
-
-#[test]
-fn global_mem_access() {
-    let global_memory = RuntimeMemoryImage::mock();
-    let mut state = State::new(
-        &variable!("RSP:8"),
-        Tid::new("func_tid"),
-        BTreeSet::from([0x2000]),
-    );
-    // global read-only address
-    let address_expr = expr!("0x1000:8");
-    assert_eq!(
-        state
-            .load_value(&address_expr, ByteSize::new(4), &global_memory)
-            .unwrap(),
-        bitvec!("0xb3b2b1b0:4").into() // note that we read in little-endian byte order
-    );
-    assert!(state
-        .write_to_address(
-            &address_expr,
-            &DataDomain::new_top(ByteSize::new(4)),
-            &global_memory
-        )
-        .is_err());
-    // global writeable address
-    let address_expr = expr!("0x2000:8");
-    assert_eq!(
-        state
-            .load_value(&address_expr, ByteSize::new(4), &global_memory)
-            .unwrap(),
-        DataDomain::new_top(ByteSize::new(4))
-    );
-    assert!(state
-        .write_to_address(&address_expr, &bitvec!("21:4").into(), &global_memory)
-        .is_ok());
-    assert_eq!(
-        state
-            .load_value(&address_expr, ByteSize::new(4), &global_memory)
-            .unwrap(),
-        bitvec!("21:4").into()
-    );
-
-    // invalid global address
-    let address_expr = expr!("0x3456:8");
-    assert!(state
-        .load_value(&address_expr, ByteSize::new(4), &global_memory)
-        .is_err());
-    assert!(state
-        .write_to_address(
-            &address_expr,
-            &DataDomain::new_top(ByteSize::new(4)),
-            &global_memory
-        )
-        .is_err());
 }
 
 /// Test that value specialization does not introduce unintended widening hints.
